# PERF: speed up CategoricalIndex.get_loc