@shenker 2018-10-19
Out of curiosity, does this fix improve `.get_loc` performance for MultiIndexes?