@topper-123 2018-10-19
All tests pass.

@jreback, I've simplified this new version a lot now by adding 'hashtable_name' and 'hashtable_dtype' to the ``dtype`` list in ``index_class_helper.pxi.in``, so we now avoid adding so many if-statements to the code.