@jbrockmendel 2018-10-19
Actually, do we have tests that cover each of the dtypes that `codes` might take?