@topper-123 2018-10-20
Currently, the algos in ``algos.pyx`` do not support uint32 and uint16. To get UInt32Engine and UInt16Engine to work we need to have at least ``ensure_uint32/_uint16`` and ``is_monotonic_uint32/uint_16``, but in that case we should probably just make all the algos have uint32/uint16 variants for consistency.

For exampel currently this fails:

```python
>>> arr = np.array([1,2,3,4], dtype=np.uint32)
>>> engine = pd._libs.index.UInt32Engine(lambda: arr, 4)
>>> engine.is_monotonic_increasing()
AttributeError: module 'pandas._libs.algos' has no attribute 'ensure_uint32'
```
As I understood @jreback, UInt(32|16)Engine is desired, so I'll add the various uint32/uint16 variants to algos.pyx to make the uint indexing engines work, is that correct?

I'll also add a test_indexing_engines.py to capture such errors at one above.