@topper-123 2018-10-20
Added various stuff to get UInt(32|16) to work and added tests for same.