@topper-123 2018-10-20
@jbrockmendel, I've made tests for ``CategoricalIndex._engine`` being of the correct type.

Also, I've made some tests for engine themselves in test_indexing_engine.py.  This last bit is needed because some engine types are not used yet in pandas. (These are ``UINT(8|16|32)Engine``), so we need to check directly that the cython methods work for those engine types.